`Refactor variable names and add noexcept to operator()`
Refactored variable names in `sort_executor.cpp` for better readability. Changed `sort_value1` and `sort_value2` to `s1` and `s2` respectively. This is a cosmetic change and does not affect code functionality.

In `aggregation_plan.h`, added `noexcept` to the `operator()` function in the `hash<bustub::AggregateKey>` struct. This indicates that the function does not throw any exceptions, potentially improving performance by allowing compiler optimizations. It also places a requirement on the function to ensure it does not throw any exceptions.